Cameras now working and effective

+++ src/org/alexdev/icarus/game/room/user/ChatType.java
@@ -14,7 +14,7 @@ public enum ChatType {
      *
      * @return the chat header
      */
-    public int getHeader() {
+    public short getHeader() {
         
         if (this == CHAT) {
             return Outgoing.ChatMessageComposer;
